Add files via upload
Create a MinHeap class as follows:
o In the constructor (initializer), create an empty list named heap and a
variable named size that is initialized to 0.
o Create the same functions as in the video demo: get_left_child,
get_right_child, get_parent, swap, remove_root
o Create an insert() function with one parameter. This function will append
the parameter to the end of the heap list and then use a similar “bubble-
up” process to move the new value to its proper position. This process
should swap two values if the current value is less than the value of the
parent of the current value.
o Create a sink_down() function with one parameter and is called from the
remove_root function. This function will use a similar algorithm to the
example in the video to reorder the nodes so that the node with the
smallest value is the root and the other nodes are arranged so that they
form a correct heap.
• Create a main program and call it at the end of the file. Within main, do the
following:
o Create an object of the MinHeap class named minheap.
o Insert the following numbers into the heap, in this order: 99, 72, 61, 58
o Print the heap in the minheap object.
o Insert 100 into the heap and print the heap again.
o Insert 75 into the heap and print the heap again.
o Call the remove_root() function and print the heap again.
o Call the remove_root() function a second time and print the heap again